# Size matters: Add command `shrink` and optimize video size output 