--- a/cmd/root.go
+++ b/cmd/root.go
@@ -6,6 +6,8 @@ import (
 	"github.com/spf13/cobra"
 )
 
+var flagUseDebug bool = false
+
 var rootCmd = &cobra.Command{
 	Use:     "sf6vid",
 	Version: "0.3.0",
@@ -20,4 +22,5 @@ func Execute() {
 }
 
 func init() {
+	rootCmd.PersistentFlags().BoolVar(&flagUseDebug, "debug", false, "More verbose logging")
 }