Closes https://github.com/techygrrrl/sf6vid/issues/15

## shrink

Adds shrink which allows the user to shrink the video frame by a percentage, e.g.

```sh
sf6vid shrink -i samples/sample-1080x720.mp4 -o samples/shrunk.mp4 -s 50
```

You'll get:

```
✅ Shrunk video was output to: /path/to/samples/shrunk_540x360.mp4
```

## censor

The censor command now outputs a suffix that indicates which side was censored and the start/end time from the source video:

```sh
sf6vid censor -i samples/sample-1080x720.mp4 -o samples/foo.mp4 --p2 --start 2s --end 9s
```

Result:

```
✅ Censored video was output to: /path/to/samples/foo_censored_p2_2s-9s.mp4
```


## trim

The trim command now outputs that it was trimmed and the start/end time from the source video:

```sh
sf6vid trim -i samples/sample-1080x720.mp4 -o samples/foo.mp4 --start 2s --end 9s
```

Output:

```
✅ Trimmed video was output to: /path/to/samples/foo_trimmed_2s-9s.mp4
```

## General quality settings

Changes the codec to use H.265 which is more optimal re: size. This change was done to all commands.

## --debug

Outputs more verbose logging if you do `--debug` with any command.